+++ src/app/api/suppliers/[id]/route.ts
@@ -64,6 +64,21 @@ export const PUT = async (req: NextRequest, { params }: { params: { id: string}}
 export const DELETE = async (_req: NextRequest, { params }: { params: { id: string}}) => {
   try {
     const { id } = params
+
+    // validate that supllier has transactions
+    const [res] = await conn.query<any[]>(`SELECT COUNT(transactionId) AS totalTransactions FROM TRANSACTIONS tra
+    INNER JOIN SUPPLIERS sup ON sup.supplierId=tra.supplierId
+    WHERE tra.supplierId = ?`, id)
+    const { totalTransactions } = res
+
+    if (totalTransactions > 0) {
+      return NextResponse.json<ApiResponse>({
+        ok: false,
+        message: 'No se puede eliminar un proveedor partícipe'
+      })
+    }
+
+    // delete supplier
     const resDB = await conn.query<OkPacket>('DELETE FROM SUPPLIERS WHERE supplierId = ?', id)
 
     if (resDB.affectedRows > 0) {
